Test: add mock tests for open api request
