+++ src/native-test/java/com/quarkus/NativeBlogResourceIT.java
@@ -1,10 +1,10 @@
 package com.quarkus;
 
-import com.quarkus.resource.BlogResourceTest;
+import com.quarkus.integration.BlogResourceITest;
 import io.quarkus.test.junit.NativeImageTest;
 
 @NativeImageTest
-public class NativeBlogResourceIT extends BlogResourceTest {
+public class NativeBlogResourceIT extends BlogResourceITest {
 
     // Execute the same tests but in native mode.
 }
