+++ src/test/java/com/quarkus/resource/BlogResourceTest.java
@@ -1,78 +1,41 @@
 package com.quarkus.resource;
 
-import com.quarkus.entity.CredentialEntity;
-import com.quarkus.entity.UserEntity;
+import com.quarkus.exception.BusinessLogicException;
 import com.quarkus.exception.ErrorMessage;
 import com.quarkus.model.Comment;
 import com.quarkus.model.Credential;
 import com.quarkus.model.Post;
 import com.quarkus.model.UserDetail;
-import com.quarkus.repository.CredentialsRepository;
-import com.quarkus.repository.UserRepository;
-import com.quarkus.security.PasswordEncoder;
-import com.quarkus.security.TokenGenerator;
-import io.quarkus.test.common.QuarkusTestResource;
-import io.quarkus.test.h2.H2DatabaseTestResource;
+import com.quarkus.service.AuthenticationService;
+import com.quarkus.service.UserService;
 import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.mockito.InjectMock;
 import io.restassured.http.ContentType;
 import io.restassured.http.Header;
-import lombok.SneakyThrows;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import javax.inject.Inject;
+import org.mockito.Mockito;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.is;
 
 @QuarkusTest
-@QuarkusTestResource(H2DatabaseTestResource.class)
 public class BlogResourceTest {
 
-    @Inject
-    UserRepository userRepository;
+    @InjectMock
+    UserService userService;
 
-    @Inject
-    CredentialsRepository credentialsRepository;
-
-    @Inject
-    PasswordEncoder passwordEncoder;
-
-    @Inject
-    TokenGenerator tokenGenerator;
-
-    private String token;
-
-    @SneakyThrows
-    @BeforeEach
-    void setUp() {
-        CredentialEntity credentialEntity = credentialsRepository.findByUsername("mrQuarkusUserName")
-                .orElseGet(() -> credentialsRepository.save(CredentialEntity
-                        .builder()
-                        .username("mrQuarkusUserName")
-                        .password(passwordEncoder.encode("password"))
-                        .build()));
-
-        userRepository.save(UserEntity.builder()
-                .name("Quarkus user")
-                .lastName("Quarkus last name")
-                .credentialEntity(credentialEntity)
-                .build());
-
-        token = tokenGenerator.generateToken("mrQuarkusUserName");
-    }
+    @InjectMock
+    AuthenticationService authenticationService;
 
     @Test
     void registerUserWithSameUsername() {
+        UserDetail userDetail = UserDetail.builder().build();
+
+        Mockito.doThrow(new BusinessLogicException(ErrorMessage.USERNAME_NOT_UNIQUE))
+                .when(userService).register(userDetail);
         given()
                 .contentType(ContentType.JSON)
-                .body(UserDetail.builder()
-                        .firstName("Quarkus user")
-                        .lastName("Quarkus last name")
-                        .username("mrQuarkusUserName")
-                        .password("password")
-                        .build())
+                .body(userDetail)
                 .when().post("/blog/register")
                 .then()
                 .statusCode(400)
@@ -80,123 +43,120 @@ public class BlogResourceTest {
     }
 
     @Test
-    void login() {
+    void registerUserWithUniqueUsername() {
+        UserDetail userDetail = UserDetail.builder().build();
+        Mockito.doNothing().when(userService).register(userDetail);
+        given()
+                .contentType(ContentType.JSON)
+                .body(userDetail)
+                .when().post("/blog/register")
+                .then()
+                .statusCode(200);
+    }
+
+    @Test
+    void loginWithTokenGeneration() {
+        Credential credential = new Credential("username", "password");
+        String token = "12345";
+        Mockito.when(authenticationService.auth(credential)).thenReturn(token);
         given()
                 .when()
                 .contentType(ContentType.JSON)
-                .body(new Credential("mrQuarkusUserName", "password"))
+                .body(credential)
                 .post("/blog/login")
                 .then()
-                .statusCode(200);
+                .statusCode(200)
+                .body(is(token));
     }
 
     @Test
-    void getUserInfoForAuthorizedUser() {
+    void getPostForUnauthorizedUser() {
         given()
-                .when()
-                .header(new Header("Authorization", "Bearer " + token))
-                .contentType(ContentType.JSON)
-                .get("blog/user")
-                .then()
-                .statusCode(200);
-    }
-
-    @Test
-    void getPostForAuthorizedUser() {
-        given()
-                .when()
-                .header(new Header("Authorization", "Bearer " + token))
+                .header(new Header("Authorization", "Bearer " + null))
                 .contentType(ContentType.JSON)
                 .pathParam("tags", 1)
-                .get("blog/posts/tags/{tags}")
+                .when().get("blog/posts/tags/{tags}")
                 .then()
-                .statusCode(200);
+                .statusCode(401);
     }
 
     @Test
-    void getPostsByTitleForAuthorizedUser() {
+    void getPostsByTitleForUnauthorizedUser() {
         given()
-                .when()
-                .header(new Header("Authorization", "Bearer " + token))
+                .header(new Header("Authorization", "Bearer " + null))
                 .contentType(ContentType.JSON)
                 .pathParam("title", "Quarkus application")
-                .get("blog/posts/title/{title}")
+                .when().get("blog/posts/title/{title}")
                 .then()
-                .statusCode(200);
+                .statusCode(401);
     }
 
     @Test
-    void getCommentsForPostForAuthorizedUser() {
+    void getCommentsForPostForUnauthorizedUser() {
         given()
-                .when()
-                .header(new Header("Authorization", "Bearer " + token))
+                .header(new Header("Authorization", "Bearer " + null))
                 .contentType(ContentType.JSON)
-                .pathParam("id", 0)
-                .get("blog/posts/{id}/comments")
+                .pathParam("id", 1)
+                .when().get("blog/posts/{id}/comments")
                 .then()
-                .statusCode(400)
-                .body(is(ErrorMessage.POST_DOESNT_EXIST));
+                .statusCode(401);
     }
 
     @Test
-    void createCommentForPostForAuthorizedUser() {
+    void createCommentForPostForUnauthorizedUser() {
         given()
-                .when()
-                .header(new Header("Authorization", "Bearer " + token))
+                .header(new Header("Authorization", "Bearer " + null))
                 .contentType(ContentType.JSON)
-                .pathParam("id", 0)
+                .pathParam("id", 1)
                 .body(Comment.builder().commentText("Quarkus comment").build())
-                .post("blog/posts/{id}/comment")
+                .when().post("blog/posts/{id}/comment")
                 .then()
-                .statusCode(400)
-                .body(is(ErrorMessage.POST_DOESNT_EXIST));
+                .statusCode(401);
     }
 
     @Test
-    void createPostForAuthorizedUser() {
+    void createPostForUnauthorizedUser() {
         given()
-                .when()
-                .header(new Header("Authorization", "Bearer " + token))
+                .header(new Header("Authorization", "Bearer " + null))
                 .contentType(ContentType.JSON)
                 .body(Post.builder()
                         .title("Quarkus Title")
                         .text("Quarkus text")
                         .tags("#quarkus")
                         .build())
-                .post("/blog/posts/post")
+                .when().post("blog/posts/post")
                 .then()
-                .statusCode(200);
+                .statusCode(401);
     }
 
     @Test
-    void getPostsForAuthorizedUser() {
+    void getPostsForUnauthorizedUser() {
         given()
-                .when()
-                .header(new Header("Authorization", "Bearer " + token))
-                .get("blog/posts")
+                .header(new Header("Authorization", "Bearer " + null))
+                .contentType(ContentType.JSON)
+                .when().get("blog/posts")
                 .then()
-                .statusCode(200);
+                .statusCode(401);
     }
 
     @Test
-    void getAllPostsForAuthorizedUser() {
+    void getAllPostsForUnauthorizedUser() {
         given()
-                .when()
-                .header(new Header("Authorization", "Bearer " + token))
-                .get("blog/posts/all")
+                .header(new Header("Authorization", "Bearer " + null))
+                .contentType(ContentType.JSON)
+                .when().get("blog/posts/all")
                 .then()
-                .statusCode(200);
+                .statusCode(401);
     }
 
-
-    @AfterEach
-    void tearDown() {
-        userRepository.findAll().forEach(userEntity -> {
-            userEntity.setCredentialEntity(null);
-            userEntity.setPosts(null);
-            userRepository.save(userEntity);
-        });
-
-        credentialsRepository.deleteAll();
+    @Test
+    void getUserInfoForUnauthorizedUser() {
+        given()
+                .when()
+                .header(new Header("Authorization", "Bearer " + null))
+                .contentType(ContentType.JSON)
+                .get("/user")
+                .then()
+                .statusCode(401);
     }
 }
